backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Document Processing
python-pptx==0.6.23
PyMuPDF==1.23.8
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# AI/ML
anthropic==0.7.8
openai==1.3.7
sentence-transformers==2.2.2
spacy==3.7.2
transformers==4.36.2
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0

# Vector Database
pinecone-client==2.2.4
chromadb==0.4.18

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0

# Data Processing
pydantic==2.5.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Email
emails==0.6

# File handling
magic==0.4.27

# Rate limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Background tasks
rq==1.15.1

# Caching
aiocache==0.12.2
